requirements.txt
fastapi
uvicorn
pydantic>=2
python-dotenv
scikit-learn
numpy
requests
langchain==1.1.3
langchain-openai==1.1.1
langgraph-cli==0.4.9
sqlalchemy==2.0.45
pymysql==1.1.2
loguru==0.7.3
# pip install --upgrade "langgraph-cli[inmem]"
langchain-community==0.4.1
pypdf==6.4.1
langchain-chroma==1.1.0